SoilSense 0.0/SoilSense/requirements.txt
flask
torch
torchvision
pillow
matplotlib
tqdm
scikit-learn
opencv-python
numpy
scipy
gradio
qrcode[pil]